GLIDE-59 #time 30m #comment Moved parsing functions to the new parser service
