+++ src/app/services/map-view.service.ts
@@ -13,6 +13,7 @@ import {getLength, getDistance} from 'ol/sphere';
 import {Circle as CircleStyle, Fill, Stroke, Style} from 'ol/style';
 import {Subject} from 'rxjs';
 import {TrackPoint} from '../../track';
+import {ParserService} from './parser.service';
 
 @Injectable({
   providedIn: 'root'
@@ -58,14 +59,8 @@ export class MapViewService {
   minDelta = Infinity;
   maxDelta = -Infinity;
   deltaRange;
-  flightDuration;
-  totalDistance;
-  e2eDistance;
-  startAltitude;
-  stopAltitude;
-  highestPoint;
 
-  constructor() { }
+  constructor(private parser: ParserService) { }
 
   initMap() {
     // Init View
@@ -147,50 +142,6 @@ export class MapViewService {
     }
     this.map.render();
   }
-
-  // Function responsible for IGC file parsing
-  parseIGCFile(filename: string, trackDay: string, callback) {
-    let trackData: TrackPoint[] = [] as any;
-    let p: string[];
-    // Accessing the file form its url
-    this.get(filename, (data) => {
-      // Separate rows and iterate through them
-      const rows = data.split('\n');
-      rows.forEach( (row) => {
-        switch (row[0]) {
-          case 'B': // B Record parsing
-            p = this.parseBrecord(row);
-            // Add the record to the trackData array
-            trackData = [...trackData,
-              {
-                Time: new Date(`${trackDay}T${p[0]}:${p[1]}:${p[2]}`),
-                Latitude: p[3],
-                Longitude: p[4],
-                Valid: p[5] === 'A',
-                Pressure_alt: parseInt(p[6], 10),
-                GPS_alt: parseInt(p[7], 10),
-                Accuracy: parseInt(p[8], 10),
-                Engine_RPM: parseInt(p[9], 10),
-              } as TrackPoint
-            ];
-            break; // TODO Add test to check length of trackData
-        }
-      });
-      callback(trackData);
-    });
-  }
-
-  // Parse a single IGC B record string into a string array
-  parseBrecord(s: string) {
-    return [
-      s.substring(1, 3), s.substring(3, 5), s.substring(5, 7), // hours, minutes, seconds
-      s.substring(7, 15), s.substring(15, 24),                 // latitude, longitude
-      s.substring(24, 25),                                     // valid
-      s.substring(25, 30), s.substring(30, 35),                // pressure alt, gps alt
-      s.substring(35, 38), s.substring(38, 42)                 // accuracy; engine rpm
-    ];
-  }
-
   // Load a track as features into to tracks layer on the map
   loadTrack(trackData) {
     const features = this.fromTrackDataToFeatures(trackData);
@@ -201,75 +152,39 @@ export class MapViewService {
   // Create feature from track data
   fromTrackDataToFeatures(trackData: TrackPoint[], dataPerFeature = 20) {
     const features = [];
-    let geometry, coords, point, alt1, alt2, t1, t2, delta;
     for (let i = 0; i < trackData.length; i += dataPerFeature - 1) {
-      geometry = new LineString([], 'XYZM');
-      for (let j = i; j < i + dataPerFeature && i < trackData.length - dataPerFeature; j++) {
-        point = trackData[j];
-        if (j === i) {
-          alt1 = point.GPS_alt;
-          t1 = point.Time;
-        } else if (j === i + dataPerFeature - 1) {
-          alt2 = point.GPS_alt;
-          t2 = point.Time;
-        }
-        // Coordinates projection from Decimal Degrees to EPSG:3857
-        coords = fromLonLat(this.fromLonLatStr([point.Longitude, point.Latitude]));
-        coords = [...coords, point.GPS_alt, point.Time.getTime() / 1000];
-        geometry.appendCoordinate(coords);
-      }
-      // delta is the steepness of the upward/downward movement for the current geometry
-      // delta > 0 means upward movement, else it's downward
-      delta = (alt2 - alt1) / this.getElapsedTime(t1, t2);
-      this.updateDeltaValues(delta);
-      features.push(new Feature({
-        geometry: geometry,
-        delta_altitude: delta
-      }));
+      features.push(
+        this.createSingleFeatureFromTrackData(trackData.slice(i, i + dataPerFeature))
+      );
     }
     return features;
   }
 
-  // Returns the total length of the track in meters
-  getTotalDistance(trackData) {
-    if (!this.totalDistance) {
-      this.totalDistance = 0;
-      let c1, c2;
-      for (let i = 0; i < trackData.length - 1; i++) {
-        c1 = this.fromLonLatStr([trackData[i].Longitude, trackData[i].Latitude]);
-        c2 = this.fromLonLatStr([trackData[i + 1].Longitude, trackData[i + 1].Latitude]);
-        this.totalDistance += getDistance(c1, c2);
+  createSingleFeatureFromTrackData(trackData) {
+    let coords, point, alt1, alt2, t1, t2, delta;
+    const geometry = new LineString([], 'XYZM');
+    for (let i = 0; i < trackData.length; i++) {
+      point = trackData[i];
+      if (i === 0) {
+        alt1 = point.GPS_alt;
+        t1 = point.Time;
+      } else if (i === trackData.length - 1) {
+        alt2 = point.GPS_alt;
+        t2 = point.Time;
       }
-
+      // Coordinates projection from Decimal Degrees to EPSG:3857
+      coords = fromLonLat([point.Longitude, point.Latitude]);
+      coords = [...coords, point.GPS_alt, point.Time.getTime() / 1000];
+      geometry.appendCoordinate(coords);
     }
-    return this.totalDistance;
-  }
-
-  // Returns the altitude in meters at the beginning of the track
-  getStartAltitude(trackData) {
-    if (!this.startAltitude) {
-      this.startAltitude = trackData[0].GPS_alt;
-    }
-    return this.startAltitude;
-  }
-
-  // Returns the altitude in meters at the end of the track
-  getStopAltitude(trackData) {
-    if (!this.stopAltitude) {
-      this.stopAltitude = trackData[trackData.length - 1].GPS_alt;
-    }
-    return this.stopAltitude;
-  }
-
-  // Returns the greatest altitude reached during the flight in meters
-  getHighestPoint(trackData) {
-    if (!this.highestPoint) {
-      this.highestPoint = 0;
-      trackData.forEach( point =>
-        this.highestPoint = point.GPS_alt > this.highestPoint ? point.GPS_alt : this.highestPoint
-      );
-    }
-    return this.highestPoint;
+    // delta is the steepness of the upward/downward movement for the current geometry
+    // delta > 0 means upward movement, else it's downward
+    delta = (alt2 - alt1) / this.parser.getElapsedTime(t1, t2);
+    this.updateDeltaValues(delta);
+    return new Feature({
+      geometry: geometry,
+      delta_altitude: delta
+    });
   }
 
   // Returns the maximum ascending speed reached during the flight
@@ -280,35 +195,6 @@ export class MapViewService {
   // Returns the maximum descending speed reached during the flight
   getMaxDescentSpeed() {
     return this.minDelta;
-  }
-
-  // Returns the time elapsed between t1 and t2 in seconds
-  getElapsedTime(t1, t2) {
-    return (t2 - t1) / 1000;
-  }
-
-  // Returns to total duration of the flight
-  getFlightDuration(trackData) {
-    if (!this.flightDuration) {
-      const t1 = trackData[0].Time;
-      const t2 = trackData[trackData.length - 1].Time;
-      const deltaT = this.getElapsedTime(t1, t2);
-      const hours = Math.floor(deltaT / 3600);
-      const minutes = Math.floor((deltaT % 3600) / 60);
-      const seconds = Math.floor(deltaT % 60);
-      this.flightDuration = `${hours}h${minutes}m${seconds}s`;
-    }
-    return this.flightDuration;
-  }
-
-  // Returns the distance in meters between the start and the end points
-  getE2EDistance(trackData) {
-    if (!this.e2eDistance) {
-      const c1 = this.fromLonLatStr([trackData[0].Longitude, trackData[0].Latitude]);
-      const c2 = this.fromLonLatStr([trackData[trackData.length - 1].Longitude, trackData[trackData.length - 1].Latitude]);
-      this.e2eDistance = getDistance(c1, c2);
-    }
-    return this.e2eDistance;
   }
 
   /*
@@ -325,15 +211,6 @@ export class MapViewService {
       this.maxDelta = delta;
     }
     this.deltaRange = this.maxDelta - this.minDelta;
-  }
-
-  // Convert parsed IGC coords ['5142113N','01751264E'] to float values in Decimal Degrees
-  fromLonLatStr([lon, lat]: string[]) {
-    const x = parseInt(lon.substr(0, 3), 10) + parseInt(lon.substr(3, 5), 10) / 60000;
-    const y = parseInt(lat.substr(0, 2), 10) + parseInt(lat.substr(2, 5), 10) / 60000;
-    const longitude = lon.substring(lon.length - 1) === 'E' ? x : -x;
-    const latitude = lat.substring(lat.length - 1) === 'N' ? y : -y;
-    return [longitude, latitude];
   }
 
   // Style function to color code upward/downward movement to each feature of the track
@@ -355,16 +232,6 @@ export class MapViewService {
   // Sigmoid function for better color range coding of the upward/downward movement
   sigmoid(x, range) {
     return 1 / ( 1 + Math.exp(-8 * x / range));
-  }
-
-  // GET html request
-  get(url, callback) {
-    const client = new XMLHttpRequest();
-    client.open('GET', url);
-    client.onload = () => {
-      callback(client.responseText);
-    };
-    client.send();
   }
 
   // Update this.infos with the given data
