+++ src/app/services/parser.service.ts
@@ -1,0 +1,153 @@
+import { Injectable } from '@angular/core';
+import {TrackPoint} from '../../track';
+import {getDistance} from 'ol/sphere';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ParserService {
+
+  flightDuration;
+  totalDistance;
+  e2eDistance;
+  startAltitude;
+  stopAltitude;
+  highestPoint;
+
+  // Function responsible for IGC file parsing
+  parseIGCFile(filename: string, trackDay: string, callback) {
+    let trackData: TrackPoint[] = [] as any;
+    let p: string[];
+    // Accessing the file form its url
+    this.get(filename, (data) => {
+      // Separate rows and iterate through them
+      const rows = data.split('\n');
+      rows.forEach( (row) => {
+        switch (row[0]) {
+          case 'B': // B Record parsing
+            p = this.parseBrecord(row);
+            // Add the record to the trackData array
+            trackData = [...trackData,
+              {
+                Time: new Date(`${trackDay}T${p[0]}:${p[1]}:${p[2]}`),
+                Latitude: this.parseCoord(p[3]),
+                Longitude: this.parseCoord(p[4]),
+                Valid: p[5] === 'A',
+                Pressure_alt: parseInt(p[6], 10),
+                GPS_alt: parseInt(p[7], 10),
+                Accuracy: parseInt(p[8], 10),
+                Engine_RPM: parseInt(p[9], 10),
+              } as TrackPoint
+            ];
+            break; // TODO Add test to check length of trackData
+        }
+      });
+      callback(trackData);
+    });
+  }
+
+  // Parse a single IGC B record string into a string array
+  parseBrecord(s: string) {
+    return [
+      s.substring(1, 3), s.substring(3, 5), s.substring(5, 7), // hours, minutes, seconds
+      s.substring(7, 15), s.substring(15, 24),                 // latitude, longitude
+      s.substring(24, 25),                                     // valid
+      s.substring(25, 30), s.substring(30, 35),                // pressure alt, gps alt
+      s.substring(35, 38), s.substring(38, 42)                 // accuracy; engine rpm
+    ];
+  }
+
+  // Convert parsed IGC coord '5142113N' or '01751264E' to float values in Decimal Degrees
+  parseCoord(value) {
+    let result;
+    if (value.length === 9) { // Longitude
+      result = parseInt(value.substr(0, 3), 10) + parseInt(value.substr(3, 5), 10) / 60000;
+      result = value.substring(value.length - 1) === 'E' ? result : -result;
+    } else { // Latitude
+      result = parseInt(value.substr(0, 2), 10) + parseInt(value.substr(2, 5), 10) / 60000;
+      result = value.substring(value.length - 1) === 'N' ? result : -result;
+    }
+    return result;
+  }
+
+  // Returns the total length of the track in meters
+  getTotalDistance(trackData) {
+    if (!this.totalDistance) {
+      this.totalDistance = 0;
+      let c1, c2;
+      for (let i = 0; i < trackData.length - 1; i++) {
+        c1 = [trackData[i].Longitude, trackData[i].Latitude];
+        c2 = [trackData[i + 1].Longitude, trackData[i + 1].Latitude];
+        this.totalDistance += getDistance(c1, c2);
+      }
+
+    }
+    return this.totalDistance;
+  }
+
+  // Returns the altitude in meters at the beginning of the track
+  getStartAltitude(trackData) {
+    if (!this.startAltitude) {
+      this.startAltitude = trackData[0].GPS_alt;
+    }
+    return this.startAltitude;
+  }
+
+  // Returns the altitude in meters at the end of the track
+  getStopAltitude(trackData) {
+    if (!this.stopAltitude) {
+      this.stopAltitude = trackData[trackData.length - 1].GPS_alt;
+    }
+    return this.stopAltitude;
+  }
+
+  // Returns the greatest altitude reached during the flight in meters
+  getHighestPoint(trackData) {
+    if (!this.highestPoint) {
+      this.highestPoint = 0;
+      trackData.forEach( point =>
+        this.highestPoint = point.GPS_alt > this.highestPoint ? point.GPS_alt : this.highestPoint
+      );
+    }
+    return this.highestPoint;
+  }
+
+  // Returns the time elapsed between t1 and t2 in seconds
+  getElapsedTime(t1, t2) {
+    return (t2 - t1) / 1000;
+  }
+
+  // Returns to total duration of the flight
+  getFlightDuration(trackData) {
+    if (!this.flightDuration) {
+      const t1 = trackData[0].Time;
+      const t2 = trackData[trackData.length - 1].Time;
+      const deltaT = this.getElapsedTime(t1, t2);
+      const hours = Math.floor(deltaT / 3600);
+      const minutes = Math.floor((deltaT % 3600) / 60);
+      const seconds = Math.floor(deltaT % 60);
+      this.flightDuration = `${hours}h${minutes}m${seconds}s`;
+    }
+    return this.flightDuration;
+  }
+
+  // Returns the distance in meters between the start and the end points
+  getE2EDistance(trackData) {
+    if (!this.e2eDistance) {
+      const c1 = [trackData[0].Longitude, trackData[0].Latitude];
+      const c2 = [trackData[trackData.length - 1].Longitude, trackData[trackData.length - 1].Latitude];
+      this.e2eDistance = getDistance(c1, c2);
+    }
+    return this.e2eDistance;
+  }
+
+  // GET html request
+  get(url, callback) {
+    const client = new XMLHttpRequest();
+    client.open('GET', url);
+    client.onload = () => {
+      callback(client.responseText);
+    };
+    client.send();
+  }
+}
